refactor: move RacingGame to Domain
