+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.controller.RacingGame;
+import racingcar.domain.RacingGame;
 
 public class Application {
     public static void main(String[] args) {
